Update: Segment tree, Special DS, Sieve

+++ code/Number Theory/SPF.cc
@@ -3,8 +3,8 @@ void buildSpf() {
   for (int i = 2; i < N; i++) {
     spf[i] = i;
   }
-  for (int i = 2; 1LL * i * i < N; i++)
+  for (ll i = 2; 1LL * i * i < N; i++)
     if (spf[i] == i)
-      for (int j = i * i; j < N; j += i)
+      for (ll j = i * i; j < N; j += i)
         if (spf[j] == j) spf[j] = i;
 }
