+++ code/Number Theory/Sieve.cc
@@ -6,7 +6,7 @@ void sieve() {
   primes.push_back(2);
   for (int i = 4; i < N; i += 2)
     notprime[i] = true;
-  for (int i = 3; i < N; i += 2) {
+  for (ll i = 3; i < N; i += 2) {
     if (!notprime[i]) {
       primes.push_back(i);
       for (ll j = 1LL * i * i; j < N; j += 2 * i)
